Fix: Same block different transaction randomness
DistributedRandomNumberGenerator does not seem to give per transaction
randomness in the same block as it is being used.

This patch adds the trigger transaction as a source of randomness.

Without this patch different transactions on the same block would result
in the same assets being sent in reply to each of the transactions.